+++ plugins/collector/api/collector.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package definecollector
+package api
 
 import (
 	"reflect"
