+++ plugins/parser/api/parser.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package deinefilter
+package api
 
 import (
 	"reflect"
@@ -24,27 +24,26 @@ import (
 	"github.com/apache/skywalking-satellite/internal/pkg/plugin"
 )
 
-//   Init()        Initiating stage: Init plugin by config
-//    ||
-//    \/
-//   Process()     Running stage:    Process the input event to convert to new event. During the processing,
-//                                   the method should also tag event type to mark the event category.
-
-// Filter is a plugin interface, that defines new pipeline filters.
-type Filter interface {
+//
+// Collector ==> RawData ==> Parser ==> SerializableEvent
+//
+// Parser is a plugin interface, that defines new Parsers for Collector plugin.
+type Parser interface {
 	plugin.Plugin
 
-	// Process produces a new event by processing incoming event.
-	Process(in event.Event) event.Event
+	// ParseBytes parse the byte buffer into events.
+	ParseBytes(bytes []byte) ([]event.SerializableEvent, error)
+
+	// ParseStr parse the string into events.
+	ParseStr(str string) ([]event.SerializableEvent, error)
 }
 
-var FilterCategory = reflect.TypeOf((*Filter)(nil)).Elem()
+var ParserCategory = reflect.TypeOf((*Parser)(nil)).Elem()
 
-// Get filter plugin.
-func GetFilter(pluginName string, config map[string]interface{}) Filter {
-	return plugin.Get(FilterCategory, pluginName, config).(Filter)
+func GetParser(pluginName string, config map[string]interface{}) Parser {
+	return plugin.Get(ParserCategory, pluginName, config).(Parser)
 }
 
 func init() {
-	plugin.AddPluginCategory(FilterCategory)
+	plugin.AddPluginCategory(ParserCategory)
 }
